pyproject.toml
[project]
name = "mcp-server-guide"
version = "0.5.0"
description = "MCP server for guidelines, project rules"
readme = "README.md"
license = {file = "LICENSE.md"}
requires-python = ">=3.13"
dependencies = [
    "click>=8.3.0",
    "mcp[cli]>=1.16.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "requests>=2.25.0",
    "watchdog>=3.0.0",
    "aiohttp>=3.9.0",
    "aiofiles>=23.2.1",
]

[project.scripts]
mcp-server-guide = "mcp_server_guide.main:cli_main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.workspace]
members = []

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_server_guide"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=src/mcp_server_guide",
    "--cov-report=term-missing",
]
filterwarnings = [
    "ignore:coroutine 'LazyPath.resolve' was never awaited:RuntimeWarning",
]

[tool.ruff]
src = ["src"]
target-version = "py313"
line-length = 120

[tool.mypy]
python_version = "3.13"
packages = ["src"]
tests = ["tests"]
strict = true

[dependency-groups]
dev = [
    "coverage>=7.10.7",
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.13.2",
    "types-aiofiles>=24.1.0.20250822",
    "types-pyyaml>=6.0.12.20250915",
    "types-requests>=2.32.4.20250913",
]
